+++ src/class/bucketingVisitor/bucketingVisitor.ts
@@ -1,6 +1,6 @@
 import { FsLogger } from '@flagship.io/js-sdk-logs';
 import { MurmurHashV3 } from 'react-native-murmurhash';
-import { FlagshipSdkConfig, IFlagshipBucketingVisitor, IFlagshipBucketing } from '../../types';
+import { FlagshipSdkConfig, IFlagshipBucketingVisitor, IFlagshipBucketing, IFsCacheManager, IFsPanicMode } from '../../types';
 import { FlagshipVisitorContext, DecisionApiCampaign, DecisionApiResponseData } from '../flagshipVisitor/types';
 import {
     BucketingVariation,
@@ -21,6 +21,8 @@ class BucketingVisitor implements IFlagshipBucketingVisitor {
 
     log: FsLogger;
 
+    cacheManager: IFsCacheManager;
+
     envId: string;
 
     config: FlagshipSdkConfig;
@@ -31,13 +33,24 @@ class BucketingVisitor implements IFlagshipBucketingVisitor {
 
     global: IFlagshipBucketing;
 
+    panic: IFsPanicMode;
+
     constructor(
         envId: string,
         visitorId: string,
         visitorContext: FlagshipVisitorContext,
         config: FlagshipSdkConfig,
-        globalBucket: IFlagshipBucketing
+        globalBucket: IFlagshipBucketing,
+        optional?: {
+            cacheManager?: IFsCacheManager | null;
+            panic?: IFsPanicMode | null; // NOTE: actually not used but must be never null if it's not the case anymore
+        }
     ) {
+        // const defaultOptionalValue = {
+        //     cacheManager: null,
+        //     panic: null
+        // };
+        // const { cacheManager, panic } = { ...defaultOptionalValue, ...optional };
         const bucketingData = globalBucket.data;
         this.config = config;
         this.visitorId = visitorId;