+++ src/config/default.ts
@@ -4,10 +4,11 @@ const defaultConfig: FlagshipSdkConfig = {
     fetchNow: true,
     activateNow: false,
     enableConsoleLogs: false,
+    enableClientCache: true, // this setting is ignored on server side.
     decisionMode: 'API',
     nodeEnv: 'production',
     flagshipApi: 'https://decision-api.flagship.io/v1/',
-    pollingInterval: null,
+    pollingInterval: null, // seconds
     apiKey: null, // TODO: remove next major release
     timeout: 2, // seconds
     initialBucketing: null,
@@ -21,6 +22,7 @@ const defaultConfig: FlagshipSdkConfig = {
 };
 
 export const internalConfig: FlagshipSdkInternalConfig = {
+    campaignNormalEndpoint: '@API_URL@@ENV_ID@/campaigns?mode=normal',
     bucketingEndpoint: 'https://cdn.flagship.io/@ENV_ID@/bucketing.json',
     apiV1: 'https://decision-api.flagship.io/v1/',
     apiV2: 'https://decision.flagship.io/v2/',